chore(tests) refactor deliverTx - claims mod
