+++ testutil/integration.go
@@ -22,12 +22,9 @@ import (
 
 	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/math"
-	"github.com/cosmos/cosmos-sdk/client/tx"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/tx/signing"
-	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	govv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -36,7 +33,6 @@ import (
 	"github.com/evmos/evmos/v11/app"
 	"github.com/evmos/evmos/v11/crypto/ethsecp256k1"
 	"github.com/evmos/evmos/v11/encoding"
-	"github.com/evmos/evmos/v11/utils"
 	evmtypes "github.com/evmos/evmos/v11/x/evm/types"
 
 	"github.com/ethereum/go-ethereum/common"
@@ -105,71 +101,6 @@ func Vote(
 
 	voteMsg := govv1beta1.NewMsgVote(accountAddress, proposalID, voteOption)
 	return DeliverTx(ctx, appEvmos, priv, voteMsg)
-}
-
-// DeliverTx delivers a tx for a given set of msgs
-func DeliverTx(
-	ctx sdk.Context,
-	appEvmos *app.Evmos,
-	priv *ethsecp256k1.PrivKey,
-	msgs ...sdk.Msg,
-) (abci.ResponseDeliverTx, error) {
-	encodingConfig := encoding.MakeConfig(app.ModuleBasics)
-	accountAddress := sdk.AccAddress(priv.PubKey().Address().Bytes())
-
-	txBuilder := encodingConfig.TxConfig.NewTxBuilder()
-
-	txBuilder.SetGasLimit(100_000_000)
-	amt, _ := sdk.NewIntFromString("1000000000000000000000")
-	txBuilder.SetFeeAmount(sdk.Coins{{Denom: utils.BaseDenom, Amount: amt}})
-	if err := txBuilder.SetMsgs(msgs...); err != nil {
-		return abci.ResponseDeliverTx{}, err
-	}
-
-	seq, err := appEvmos.AccountKeeper.GetSequence(ctx, accountAddress)
-	if err != nil {
-		return abci.ResponseDeliverTx{}, err
-	}
-
-	// First round: we gather all the signer infos. We use the "set empty
-	// signature" hack to do that.
-	sigV2 := signing.SignatureV2{
-		PubKey: priv.PubKey(),
-		Data: &signing.SingleSignatureData{
-			SignMode:  encodingConfig.TxConfig.SignModeHandler().DefaultMode(),
-			Signature: nil,
-		},
-		Sequence: seq,
-	}
-
-	sigsV2 := []signing.SignatureV2{sigV2}
-
-	if err := txBuilder.SetSignatures(sigsV2...); err != nil {
-		return abci.ResponseDeliverTx{}, err
-	}
-
-	// Second round: all signer infos are set, so each signer can sign.
-	accNumber := appEvmos.AccountKeeper.GetAccount(ctx, accountAddress).GetAccountNumber()
-	signerData := authsigning.SignerData{
-		ChainID:       ctx.ChainID(),
-		AccountNumber: accNumber,
-		Sequence:      seq,
-	}
-	sigV2, err = tx.SignWithPrivKey(
-		encodingConfig.TxConfig.SignModeHandler().DefaultMode(), signerData,
-		txBuilder, priv, encodingConfig.TxConfig,
-		seq,
-	)
-	if err != nil {
-		return abci.ResponseDeliverTx{}, err
-	}
-
-	sigsV2 = []signing.SignatureV2{sigV2}
-	if err = txBuilder.SetSignatures(sigsV2...); err != nil {
-		return abci.ResponseDeliverTx{}, err
-	}
-
-	return BroadcastTxBytes(appEvmos, encodingConfig.TxConfig.TxEncoder(), txBuilder.GetTx())
 }
 
 // DeliverEthTx generates and broadcasts a Cosmos Tx populated with MsgEthereumTx messages.
